% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plsr.R
\name{generate_plsr}
\alias{generate_plsr}
\title{Identify significant PLSR coefficients for pairs of left and right objects.}
\usage{
generate_plsr(left, right, p_val_threshold = 1e-04, verbose = FALSE)
}
\arguments{
\item{left}{}

\item{right}{}

\item{p_val_threshold}{}

\item{verbose}{Whether to print output.}
}
\value{
A list containing the following named values:
p_list - A list containing left-right pairs with significant values (above p_val_threshold)
non_p_list - A list containing left-right pairs with non-significant values
comp - the maximum number of PLS components.
}
\description{
This function computes PLSR coefficients and p-values between the rows of the
two matrices provided ("left" and "right"). This regression is performed separately
for each left item; i.e. in each iteration, a separate PLS regression is performed
which explains
}
\details{
The maximum number of components for PLSR is the number of columns in the right
matrix or the number of rows minus one, whichever is less.
}
