% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3DNet.R
\name{kneedle}
\alias{kneedle}
\title{Find the knee/elbow point in a vector using the Kneedle algorithm.}
\usage{
kneedle(values, sign)
}
\arguments{
\item{values}{The values to find a knee/elbow in.}

\item{sign}{-1 if the values are mostly decreasing, 1 if the
values are mostly decreasing.}
}
\value{
The index of the knee/elbow.
}
\description{
This function uses the Kneedle algorithm (Satopaa 2011)
to find the index of the knee point in the provided vector.
If the values are mostly increasing, use sign = 1. If they are
mostly decreasing, use sign = -1.
}
