% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3DNet.R
\name{max_silhouette_consensus}
\alias{max_silhouette_consensus}
\title{Pick k based on the maximum silhouette score}
\usage{
max_silhouette_consensus(data, k_range)
}
\arguments{
\item{data}{The data with an unknown number of clusters.}

\item{k_range}{The range of possible k values}
}
\value{
The chosen value of k.
}
\description{
This function uses the average silhouette width (Rousseeuw 1987),
as implemented in the NMF library (Gaujoux 2010).
}
