% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3DNet.R
\name{max_ward_kl}
\alias{max_ward_kl}
\title{Pick k based on the maximum KL index of Ward clusterings}
\usage{
max_ward_kl(data, k_range)
}
\arguments{
\item{data}{The data with an unknown number of clusters.}

\item{k_range}{The range of possible k values}
}
\description{
This function uses the NbClust package to
}
