% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nmf_modules.R
\name{normalize_nmf}
\alias{normalize_nmf}
\title{Normalize NMF}
\usage{
normalize_nmf(left_data)
}
\arguments{
\item{left_data}{The data to be normalized}
}
\description{
This function applies an antilog transformation to the data,
ensuring that all values are positive, and then scales the
result from 0 to 1.
}
