% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plsr_60_fc.R
\name{plot_plsr_result}
\alias{plot_plsr_result}
\title{Plot PLSR result}
\usage{
plot_plsr_result(plsr_result, path = ".")
}
\description{
This function generates plots that demonstrate how well (or poorly)
the PLSR models fit.
}
