% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nmf_modules.R
\name{user_pick_k}
\alias{user_pick_k}
\title{User picks k}
\usage{
user_pick_k(nmf_data, k_range)
}
\arguments{
\item{nmf_data}{The data whose number of clusters is in question}

\item{k_range}{The numbers of clusters (k values) being considered.}
}
\description{
This is a k_picker callback that simply plots the performance
of NMF on every number of clusters in k_range, and asks the
user for input.
}
