% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3DNet.R
\name{write_all_to_disk}
\alias{write_all_to_disk}
\title{Write a nested list as nested folders, with data.frame objects as csvs.}
\usage{
write_all_to_disk(name, object, folder)
}
\arguments{
\item{name}{The name of the top-level folder to be written}

\item{object}{The nested list}

\item{folder}{The parent folder where the top-level folder should be placed.}
}
\description{
Write a nested list as nested folders, with data.frame objects as csvs.
}
\examples{
# For example, if the object is
object = list(
    item1 = list(
        df1 = data.frame(a=c(1,2),b=c(3,4)),
        df2 = data.frame(a=c(1,2),b=c(3,4))
    ),
    item2 = list(
        df3=data.frame(a=c(1,2),b=c(3,4))
    ),
    df4 = data.frame(a=c(1,2),b=c(3,4))
)
# and the call is
write_all_to_disk("toplevelfolder", object, ".")
# The following files and folders would be written:
# toplevelfolder/
#     item1/
#         df1.csv
#         df2.csv
#     item2/
#         df3.csv
#     df4.csv
#

}
