% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen3DNet.R
\name{create_gen3DNet}
\alias{create_gen3DNet}
\title{Create a general 3D network.}
\usage{
create_gen3DNet(left, right, nmf_nrun = 10, k = c(),
  k_picker = max_ward_kl, seed = 123456, p_val_threshold = 1e-04,
  out_folder = "gen3DNet", verbose = TRUE)
}
\description{
This function performs calculations to generate a 3D network.
It is not meant to be called directly. You should use gen3DNet
instead. See help(gen3DNet) for further documentation.
}
